Update patch set 1
Patch Set 1:

I tested it by decompiling the apk mentioned above to find out where it downloaded the list and dictionaries from. Turned out it's hosted on www.gstatic.com so I downloaded the metadata.json it downloads from there, added an entry for my dictionary and then used a simple hosts entry to redirect it to my own host. This installs the dictionary as long as the matching dictionary is NOT included in LatinIME.apk.

Judging by the comments in the commits below it should fallback on the main dicitonaries, ie. the ones included in the apk. I haven't been able to make it do that though.

http://bit.ly/w2iNxz
http://bit.ly/wpsfw7

Patch-set: 1