Update patch set 2
Patch Set 2:

This is only part of the fix, the rest of the fix is a 270degree rotated version of the default CM animation sized appropriately. I can hold off until there is a new version and resubmit as cm_rotated270.zip to be housed here rather than "local" as that seemed objectionable.

Patch-set: 2