Update patch set 1
Patch Set 1:

Yes, I have of course used the new makedict to create it and it works fine if I trick LatinImeDictionaryPack.apk to install it. It will not get picked up LatinIME though. I just that I perhaps need to clear the data for it.

See http://review.cyanogenmod.com/#change,11505 for more information.

Patch-set: 1