Update patch set 2
Patch Set 2:

On the p5/p5wifi (which is a 1280x800 landscape device) I added the following to p5-common.mk and currently use the Xoom ICS bootanimation.

PRODUCT_COPY_FILES += \
     $(LOCAL_PATH)/bootanimation.zip:system/media/bootanimation.zip

Seems to overwrite the default one without needing any changes in vendor/cm.

Works fine for me but in the future I suppose we will need both portrait and landscape bootanimations in a bunch of different sizes.

Patch-set: 2