Update patch set 3
Patch Set 3:

@Steve: That would be the SGH-i957 I guess? Here in Sweden the Qualcomm (LTE) variants are called P7320. That's propably the modelname used in Europe.

Anyway, upgraded my P7310 from official 3.1 to official 3.2  the other day and the 90 degree compass problem is actually in the officiall 3.2 too.

The wrongly oriented HDMI is most likely due to the fact that we use Ventana proprietary libs that are for a landscape oriented display while the 8.9 (non LTE variants at least) has a portrait oriented display.

Apart from those issues I think it shares all the know issues with the p4 builds. Camera is obvious as we do not have the libs and voice calls isn't supported in any stock ROMs. Not sure why the screen rotation "blacks out" for halv a second but as I understand the p4 builds to the same thing.

Anything missing that needs to be done to get this one merged?

Patch-set: 3