Update patch set 1
Patch Set 1:

Ah, yes I remeber reading that thread back then when you posted it. I actually posted in it.

It's nice to see more dictionaries with frequency information. I'm working on a Swedish one. Currently it is based on wikipedia though but perhaps a forum is more fitting if filtered enough.

Patch-set: 1