Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I have an idea for the resolution.

We can add a new define to the device's cm.mk which states the resolution
Something like 
BOOTANIMATION_RESOLUTION := 1280 720
And in vendor/cm/prebuilt/bootanimation, there can be folder for various resolutions (drawable-hdpi and so on)
Then a script would check for the resolution, and echo to it to desc.txt, and the folder to use according to the resolution.
And maybe even resize it by using `convert`

Patch-set: 1