Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

The horizontal 1280x800 version is a zip within a zip and needs to be unzipped (once) first in order to work on my Samsung Galaxy Tab 8.9.

Also it restarted the animation from the begging and ran halfway through a second time before it was fully booted.

Patch-set: 1