Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

No, you forgot to change the Change-Id line in your commit message so it ended up like a new commit instead of a patchset.

It's probably best to revert back to the original commit below, remove the DS_store file and then git --amend it (and verify that the Change-Id is the same as the one in the original).

http://review.cyanogenmod.com/#change,13978

Patch-set: 1