Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

We've already got unofficial builds working for various 7227 devices from ZTE so it seems like its time to go official.

Do we really need to hold off just so video's play a bit smoother?

Or is it a question of wanting less supported devices to have to get OpenMAX working on (when we have the NDK components) for now?

Patch-set: 1