Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

I would be against the removal of the "update-" prefix & "-signed" suffix. It may just the Android old timer in me, but the removal of this just feels wrong.

Also the addition of "EXPERIMENTAL" is redundant when we already have SNAPSHOT...

Patch-set: 3