Update patch set 1
Patch Set 1:

Was this done by stretching the current 720x1280 bootanim?

I'm wondering if there's a way to request a re-render from the original author.

On the other hand I assume this will eventually be replaced by a Cid bootanim so it probably doesn't matter too much if the current version isn't perfect, as long as it's there as a placeholder for when the new bootanims are rendered.

Patch-set: 1