Update patch set 3
Patch Set 3:

sorry for being off topic! I always wondered if I could force the re-creation of build.prop somehow. It happens that I change just a small thing and do a rebuild with the otherwise untouched sources. Then the output file gets the date from whenever the last full build was. I even started to manually delete buiild.prop in /out befure the build. Can't this be changed so we really have the build date in final update file and not the date when build.prop was first created?

Patch-set: 3