Update patch set 1
Patch Set 1:

It is automatic-start-only-if-somebody-actively-creates-/data/droppear/.ssh/authorized_keys, which is at least some kind of surprise protection.

Additional Settings support would be great - in my opinion - but I don´t know whether it would be regarded as too much a non-smartphone thing to make it.

I imagine a separate settings screen with least something like this:
 * checkbox "enable ssh daemon" to generally enable (opt in)  the functionality (other settings grayed out when unset)
 * toggle / switch "Start ssh daemon" / " Stop ssh daemon" representing the de facto running state and permitting to change it - would be nice to have a corresponding status bar and profile toggle, too
 * checkbox "start on boot"

Elaborating, I´d also love to have some kind of user account editor, managing a passwd/shadow/group file or something similar, ideally also operational for the Terminal app and/or ADB. Regarding groups a tie in with the permissions system would be good, resulting in user accounts belonging to relevant groups or not. Regarding ssh support part of the account settings would / could then be an manage/import pubkey function. I think this could tie in nicely with incoming-ssh-to-rsync usage, making it possible to restrict rsyncing to a specific user role / that ones access rights.

I´m still no good at Java / App programming, unfortunately, but I could tackle the neccessary changes to the dropbear sources and corresponding C library support. So... somebody wants to collaborate on making this happen?

Patch-set: 1