Update patch set 4
Patch Set 4:

Are we limiting builds to once per day (for non-NIGHLTY)? If not, then I would suggest being able to use a timestamp instead of a datestamp.

I'd also like to point out that this datestamp could be misleading if someone were to build without a current reposync. I guess we can assume that the "official" builds will be based on a current reposync, but what about the "unofficial" builds. What happens when someone builds from source that's 30 days old? Should that build be stamped with today's date in the version? Given this, I think the datestamp for "unofficial" builds should be "opt-in".

I've tried using this and having my local builds be datestamped in UTC is very annoying. 

And most important, I'd really like to have the ability to inject my own timestamp/tag (without having a datestamp already there). I usually do many builds per day and a datestamp just doesn't cut it. I tag my builds with the same tag that I save the log output of the build and the manifest (repo manifest -r). To manage the output of the builds, if is very useful to be able to know the ROM filename before the build starts. With this proposed change, you cannot. This could be a new buildtype or
one of the existing ones. I don't care which.

I have uploaded a patchset
 ( http://review.cyanogenmod.com/16744 )
which (I think) addresses all of the above (and a little more). However, if it is felt that it is not suitable, please give me SOME OPTION to NOT have the datestamp injected into MY BUILDS so that I can inject my own
timestamp/tag/etc. Oh, and don't tell me I can always modify the build system after every reposync.

Patch-set: 4