Update patch set 1
Patch Set 1:

Hmm...  This one is less clear in that regard.  It belongs to the same historical class of devices as the other aries variants, which have been in the repos by name for a long time.  On the other hand, since Samsung is getting really bad about device names by giving multiple completely different devices the same or similar names (galaxys2att was renamed to i777 due to frequent shitrocket confusion...), going forward we really need to start using model number.

I guess the question is - for this one - old convention (because it's an older device) or new convention (because it's a newly supported device)?

Patch-set: 1