Update patch set 1
Patch Set 1:

Following with the same logic as the i777... I'd go with renaming.

There are no CM9 builds out there (official ones at least) with the market-name naming. On top of that, we've recently made it a guideline that for consistency's sake, all devices from the same vendor should use the same naming scheme.

Patch-set: 1