Update patch set 1
Patch Set 1:

I think that for consistency with the device's "close relatives" - this should be allowed to remain as is.

In my opinion, the real driver for moving to the model number based scheme was the total mess that Samsung created with the Galaxy S II.  The original Galaxy S line is clearly distinguishable by their marketing names, and again - I think fascinate should be allowed to be "grandfathered in" under the old naming scheme due to its position as a variant of the other devices.

However future devices should definitely go by model number.

Even boardname isn't going to work well - see Samsung renaming C1 to U1, and the upcoming "p4note" which will make pershoot's head explode.  God forbid Samsung releases a carrier variant of that one...

Patch-set: 1