Update patch set 6
Patch Set 6:

why is BUILDTYPE set to EXPERIMENTAL if CM_EXTRAVERSION is set?

surely that's going to make all kangs EXPERIMENTAL and make UNOFFICIAL a bit pointless  (as most kangers would put their name as CM_EXTRAVERSION to help differentiate).

and it makes it seem like a kang is some sort of official but experimental build.

Patch-set: 6