Update patch set 1
Patch Set 1:

I suppose we pretty much have to require a reboot for this to work anyway. A better way would be to use init's new property lookup stuff to write the symlink. I really want to avoid using this init.d scripting stuff for anything important.

Writing a doc on this stuff now

Patch-set: 1