Update patch set 6
Patch Set 6:

Ricardo, Yes, they will. Sorry if it wasn't clear by the context, but I'm talking about the names generated by this patch. Using your scenario (and this patch), the filenames would be exactly the same and you would have no way to tell between the two (which was the point I was trying to make :)).

Patch-set: 6