Update patch set 1
Patch Set 1:

I understand your point... And you are right we have no real answer on whether camera will ever function, however with the "rumored" G-Slate 2 with almost identical HW and ICS, we may end up being able to hack the cam into function. I discussed this with Steve a little bit in -priv and he was the one who told me he has p4 in nightlies w/no cam so I figured why not re visit the idea. I am not sure what the status on p4 and the cam are, he did say he hacked it to load but it wasn't "functional" so I figured at this point we are in the same boat. Correct me if I am wrong but this was where I was at so that is why I bumped this again.

Patch-set: 1