Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Interesting. Are you sure the dictionary format has changed though? Bigram dictionaries were added as part of the changed format for ICS. It seems everything they added in JB with regards to prediction could be done with the bigram statistics. Unfortunately there is no sample or description of the format so we can't create custom dictionaries using it.

Thanks for adding me. The commit looks good to me.

Patch-set: 1