Update patch set 1
Patch Set 1:

Björn, I'm not sure how much the format has changed. According to the makedict information, a type of "2" means it was made for JellyBean, so I assumed that meant the format changed. This is my first look into the dictionaries, so I'm not familiar with the previous versions.

The good news is, you can use makedict to decompile these ones and could probably figure out the format.

Patch-set: 1