Update patch set 1
Patch Set 1:

Pavel,

I'm not sure where the new wordlists are coming from, but I am fine with that. Please ensure that the updated English dictionary is used though. This larger one from the LatinIME keyboard includes the required items for sentence prediction. The default AOSP dictionaries do not have this.

Patch-set: 1