Update patch set 1
Patch Set 1:

I believe the app for downloading the additional language packs is floating around in a newer JB gapps package. Looking at the version information, it states that the english dictionary is newer. However, I am not able to decompile the dictionaries on that site with makedict, so I can look into it.

Patch-set: 1