Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

I have fixed the error with _el that Pavel mentioned and have included the commit he linked to.

I attempted to include the dictionaries that Björn linked to, but they do not work with the LatinIME keyboard. I'm assuming those require the DictionaryPack app to be used.

Patch-set: 2