Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

Hrmm... I went back to a stock JB build on maguro and it looks like English can download properly on it. I'm going to investigate that a bit later and see if we can fix the downloading ability. If so, I'll move to have all dictionaries removed from our overlays instead.

Patch-set: 4