Update patch set 4
Patch Set 4:

I agree. The dictionary overlay in its current state just wastes space and makes LatinIME.apk big, stuffing it with dictionaries that can't be used, because they have wrong names and are in wrong folders. Something should be done about it asap.

Brint, are you OK with me uploading new patchset with removed EN dictionary (so that the one that is already part of LatinIME resources is used)?

The 4MB EN dictionary can be then uploaded as a separate commit.

Patch-set: 4