Update patch set 4
Patch Set 4:

Brint: Uhm, the reason the dictionaries are included like this is because most of them are not official ones and can therefor not be downloaded like that. LatinIME in general has pretty poor dictionary support if you remove the unofficial ones. It's also normal to include some of the most common ones as a fallback.

I did notice a strange behaviour before this patch though. When I selected Swedish it would still auto-correct my Swedish words using the English dictionary. Let's just say it wasn't great in terms of usability. ;)

Patch-set: 4