Update patch set 4
Patch Set 4: Looks good to me, but someone else must approve

In that case, I would prefer the patch to go in as-is so that the English language is fully capable as well. Once downloading is figured out, we can submit a future patch to remove those languages that can be pulled (including English). In the meantime, I don't feel that the 3MB difference is is a problem, given the benefit it provides to those using the English input.

Patch-set: 4