Update patch set 2
Patch Set 2:

This was copypasta'd from cm7. I added it for the mesmerize, which is just a clone of the fascinate with different apns. Mainly added it here to cater to users that were using fascinate builds on their mesmerize.

Never seemed to cause issues with cm7. That said, I don't particularly care if it it is reverted.

Patch-set: 2