Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

310120 is Sprint's mcc/mnc.  Adding this is quite likely to break MMS support on all Sprint devices.

USCC shouldn't be using 310120.  We've been trying to figure out what USCC acutally _does_ use, and from the USCC S3's stock ROM (?) it seems to be 311220, which is what we're going to soon be using in Jellybean [1].

Which devices are you looking to fix here?  If the device contains a hardcoded mcc/mnc pair, they will have to be adjusted.

[1] http://r.cyanogenmod.com/#/c/21318/

Patch-set: 1