Update patch set 1
Patch Set 1:

Im using this fine on a Virgin mobile device. It also fixed mms for a us cellular device (the marvelc and was pulled from a stock sense rom)  This being said, it still does need testing on a sense phone. but at the same time worked fine on the USCC phone my tester had

Patch-set: 1