Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

requires jellybean branch of : https://github.com/teamhacksung/android_device_samsung_fascinatemtd

Device is up-to-speed as the rest of the SGS1 devices in terms of functionality, and shares most of their common code. No additional modifications to CM are needed to build or use it.

Patch-set: 1