Update patch set 2
Patch Set 2:

Has anyone else tested this with a horizontal device? Not sure if I'm doing anything wrong or not.

If I open the zip, I see the square boot animation rotated 90 degree right. But if I open the old horizontal zip, I see the boot animation not rotated.

Patch-set: 2