Update patch set 2
Patch Set 2:

> Then just change it in your cm.mk file to use the verticle-800.zip.

That would not make any sense. The tablet is a horizontal device and the default specifies that it use a horizontal bootanimation, which is correct. Using the vertical one only works because this bootanimation has a square shape. And I can't do that because I need the horizontal-1200.zip

Patch-set: 2