Update patch set 2
Patch Set 2:

How is the pile of cross-repo dependencies driven by this change going to be handled?  A large number of device trees need to be updated at the same time this gets merged, or the builds will break.  Since the new names aren't supported currently, the device trees can't be merged in advance.

For example, given the typical merge latency of grouper, that device is going to be broken for at least 1-2 weeks...

Patch-set: 2