Update patch set 2
Patch Set 2:

Let's calculate this:
How many devices are there? ~80
How many are active? ~30
Lets say that each takes about MAX 5 min to change and commit
So that would be 5x30=150 min MAX
Which is 2hrs and 30min
I honestly don't mind updating all the devices this weekend as I have tons of time, but I don't have access to push to any repo... And it shouldn't take longer than 2 min using the built in editor in GitHub

Patch-set: 2