Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

Since it seems like the consensus was to not separate horizontal + vertical, I did that to move thing along. The 800.zip still is the same wrong resolution of 720^2.

The problem isn't so much the time it takes to modify them -- can do that with find + xargs + sed, the issue is the need to merge all of them. For the short run it might make sense to add a bit of logic to common.mk to deal with devices specifying the old bootanimations.

Patch-set: 3