Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

Missing 1200

One of the reason against merging vertical and horizontal despite that they are the same is that if we decide to go for non square boot animation for CM 11, we won't need to change every device again.

Of course, we can also try to build a selection system where the device specifies a target resolution and the filename is automatically chosen.

Patch-set: 3