Update patch set 4
Patch Set 4:

I'll test the change tomorrow.

The problem isn't making the change for each device - the problem is getting 30-40 devices all merged at the same time this patch does.

If gerrit had a concept of cross-repo dependencies it would be easy - +2 all of the devices before setting this one to merge, and make em' all depend on this - but it's not possible to do cross-repo dependencies.  :(

Patch-set: 4