Update patch set 4
Patch Set 4:

Andrew, grouper has a maintainer, nemith, and he has +2 as well on grouper repos.

Also, i believe the folks at Google are working on cross-project dependencies.

An ideal way would be to +1 all the device changes, and this when it's ready, and then ask someone from the global +2 to merge em all at once.

Patch-set: 4