Update patch set 5
Patch Set 5: Looks good to me, but someone else must approve

It doesn't have to have a zip. It just needs to be configured to use the 720x one. If they are gonna be the exact same thing why just duplicate them? Leave one, name it vertical-720+ and let people use that

Patch-set: 5