Update patch set 5
Patch Set 5:

I don't like the way the check is done at the moment, mangling the old file names. We might as well just make it flexible and let the target specify the width and height of their screen, then we pick the best bootanimation to use from that.

Patch-set: 5