Update patch set 1
Patch Set 1:

As pngcrush can optimize png files directly since version  1.7.22 (which was released December 2011) it would be better to use the -ow (overwrite) option. This way we wouldn't need the mv anymore and could save I/O.

The problem with the overwrite option would be that people with not up to date buildservers would not get optimized png files.

Patch-set: 1