Update patch set 1
Patch Set 1:

I considered this, and decided against it. Switching from self-built unofficial builds to nightlies (or stable) works fine using CM Updater (I tried it a number of times), and it's something we'd like to keep. For the self-built crowd, it's a harmless option with an easy path back to "real" CM builds.

I understand that it doesn't make sense to have this on non-supported devices, but that's a whole other ball of wax, and we have no reliable way of distinguishing them at build time.

Patch-set: 1