Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

i'd personally really appreciate if this could be made to work. i have already tried this mod myself and it does bring up the settings menu for it but the menu is empty.  

Perhaps they were moved to a different area or could be ported from ICS?

Patch-set: 2