Update patch set 1
Patch Set 1:

OK, from Tortel's logs it appears that when LTE isn't working, it's because the Sprint MMS apn was "somehow" previously set as the preferred APN.  All data call attempts are made using that apn, which fail, so a data session can never be established.

Now, there's code in TelephonyProvider to determine when apns-conf.xml is changed, and when it is, to invalidate the previous preferred APN (since the list is no longer valid) and reattempt data sessions using all apns for that carrier.  The MMS apn is used first since it's earliest in the list, and since the data call setup fails, the LTE apn is selected next, which connects and the new preference is set.

So, that's why this patch _appears_ to fix the problem.  It's not the addition of new apns, any changes made to apns-conf.xml will force the above process which _usually_ results in LTE working.

How does LTE break then?  I don't know, but I suspect it's when CM updates push a new apns-conf.xml, occasionally the Sprint MMS apn "successfully connects", possibly the result of a race with the teardown of the CDMA data connection.  When that happens, it's set as a preference, and LTE doesn't work again until the file is modified again.

Probably the fix here is to modify the Sprint MMS apn to be used for mms only.  I'm not entirely sure why we have it in the first place, but I'll play around with it and see if something reasonable can be done there.

Tortel/Dees: Would you mind modifying "apns-conf.xml" and rebooting a few times?  Just change the name of one of the non-Sprint carriers, it just needs to change the CRC32 checksum of the file.

I'd like to get a radio logcat, if possible, of when an apns-conf.xml update breaks LTE.  That would confirm my suspicion above.

Patch-set: 1