Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I have to agree here with not uncommenting these APNs. They are duplicates of the APNs below them (as mkasick pointed out). The only difference is the carrier name. The only reason the commented and the test APNs above were left in is because we wanted to show that those APNs were programmatically removed or existed period (for reference only). If anything the test and commented APNs can be removed and the FOTA APNs below should get the carrier name from them for the sake of naming only. The carrier name does not matter however at all. It's just a display name for the APN.

The MMS APN however with default, dun, fota, or anything else shouldn't matter here either. When the telephony stack looks for a eHRPD or LTE APN it looks for a APN with the corresponding bearer (13 = eHRPD and 14 = LTE). If a APN with the needed bearer doesn't exist, then it will bail for lack of a suitable APN for LTE or eHRPD. It shouldn't attempt to use another APN that doesn't have a suitable bearer for the radio tech.

I don't really understand why it's failing to connect. But I can say for sure it's not the APNs. The real issue is deeper and resides somewhere else in the telephony stack.

No LTE or eHRPD here so full logs of system and radio are needed. I wouldn't post them here for privacy purposes but feel free to email them.

Patch-set: 1