Update patch set 1
Patch Set 1:

On a more intriguing question. Why is changing the fota APN having any affect on data at all? The fota APN is for iOTA only and is not needed to acquire a data connection. It's only used to update the provisioning info on the (e)UICC. Our devices are already provisioned for the network and do not need provisioning info updated via OTASP for NV or iOTA for (e)UICC. In the EFS you can see the only APN needed and allowed for profile0 is "n.ispsn", and the iOTA APN is allowed through since it's defined in the fallback APN list. But by no means is  iOTA required to acquire LTE or eHRPD data connections and only "n.ispsn" (internet APN) is.

We would only require a fota APN if we needed to provision the (e)UICC. Which isn't the case here at all since we have already provisioned it in the stock build

Patch-set: 1