Update patch set 1
Patch Set 1:

> It shouldn't attempt to use another APN that doesn't have a suitable bearer for the radio tech.

GsmDataConnectionTracker::buildWaitingApns accepts apns with the condition "(apn.bearer == 0 || apn.bearer == radioTech)".  So yes, having a carrier-matched "default"-type apn with bearer 0 will be used for LTE attempts.  We've also seen this in Tortel's logs.

> I don't really understand why it's failing to connect.

It fails to connect because, "somehow", the MMS apn gets selected as the preferred apn, which I think is the result of a one-time "successful" connection (perhaps just a race condition with the closing CDMA connection).

Honestly though, it shouldn't attempt to connect with the MMS apn to begin with.  If we could prevent it from doing that, the problem would be solved.

> Why is changing the fota APN having any affect on data at all?

It changes the CRC32 checksum on apns-conf.xml, which forces the next LTE connection attempt to use _all_ suitable apn types, not just the one that "worked" previously (which, in the LTE-broken case is the MMS apn).

It seems to me there's two possibilities for fixes:

- Modify code as to reject the MMS apn when attempting an LTE connection.  May introduce compatibility issues with other carriers, i.e., I don't know if it's typical for "GSM" LTE providers to have serarate LTE apns.

- Modify code to preference bearer-specific apns when attempting an LTE connection.  May not completely eliminate the problem, but should signifcantly reduce the chance of it happening.

- Modify the apns to ensure the MMS apn isn't used for LTE connection attempts.  Whether that's remove "default" from it's type, remove the entry entirely, etc., I don't know.

Personally I'm interested in an apn-focused approach as it doesn't require us to maintain any additional code modifications against AOSP.  But if that's not an appropriate thing to do, then we'll have to dig into the code.

Patch-set: 1