Update patch set 1
Patch Set 1:

> buildWaitingApns: Preferred APN:310120:310120:[ApnSettingV2]

This line indicates that you already have a preferred apn set from a previous "successful" connection.  If you modify /system/etc/apns-conf.xml in such a way as to change the file checksum (modify a non-Sprint entry will do), you'll likely see the same behavior as we do in your radio logs.  Which is that the first eHRPD/LTE connection attempt actually uses the mms apn.

> the MMS APN should have no affect on eHRPD and should never be used to even attempt connection when a APN is present with the required bearer.

This may be true, however, that's not what the code does.

Patch-set: 1