Update patch set 1
Patch Set 1:

Gotcha.
The removal of the minor/maintenance version was also fully intentional for everything that isn't a release, and we definitely don't want it in nightlies nor self-builds. That creates too much version-derived confusion, and it was one of the main motivators for trimming it down.

Experimentals might arguably use it, but on the other hand they can take advantage of EXTRAVERSION for unicity, so there isn't much point there either.

Patch-set: 1