Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

opticharger and squisher save me about 20mb in /system (ok the zip is only about 2mb smaller) and takes no longer during the build.

can we make this optional perhaps - like enabling it in common_mini_phone.mk or something?

the alternative is to optimise all the png's in cm10.1 and re-upload, although that won't help with the resources/apk's handled by zipalign and aapt.

Patch-set: 3