Update patch set 2
Patch Set 2:

@Steve -

It is varying from host to host for some unknown reason. Currently, we get nothing out of supporting other arches, and we ARE breaking existing devices, so for time being would like to revert it while we figure out what is going on with precedence.

@Brandon -

I will have to look into BUILD_PREBUILT.

Patch-set: 2