Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

But should the changelog contain the changes between CM versions (besides updating the Android version) or focussing on the extras CM provides on top of AOSP?
E.g. DSP Manager was already in Gingerbread, and I don't think that should be in the changelog for cm-10.1.

Patch-set: 1