Update patch set 1
Patch Set 1:

Marco: I see what you mean but I disagree. The thing is that the latest few releases have not been simple merges of changes but rather manual porting of features and due to framework changes etc. between versions, you can't assume a feature is included in the new release just because it existed in the previous one.

Patch-set: 1