Update patch set 1
Patch Set 1:

As there was no Changelog for CM 9/10, we are lacking recent precedence here, but my thought is as follows:

There is no guaranteeing that a previously supported feature in CM 7 will make its way to the next CM major cycle. As such, there is a benefit to explicitly calling out the feature as a change item for the new CM major version, in this case 10.1. 

I would agree if the item was listed in the CM 10.0 changelog, in which case the item is not a true 'change' in the 10.1 code branch. 

The inverse of your reasoning would mean that we would add "Removed: Permissions management" as that item is not available in 10.1, but was in CM7.

Patch-set: 1