Update patch set 1
Patch Set 1:

No, it doesn't mean that. I'll try to explain me better.

On fresh install (full wipe) or when a new user is created, SetupWizard from gapps change the value of Settings.Secure.USER_SETUP_COMPLETE when wizard is complete. When gapps is not present in the device, Settings.Secure.USER_SETUP_COMPLETE is never changed.

The consecuence of this:

* on tables the quick settings panel is disabled while Settings.Secure.USER_SETUP_COMPLETE is not set. This mean that the user only can pull down the notifications panel, but not the quick settings panel. On phones this notifications and settings share the same panel, that is always enabled.
* keyguardselector is also disabled by this setting.

Patch-set: 1