Update patch set 1
Patch Set 1:

You are correct I have no clue what you are saying.  "IPM"?

IDK, it works.. Unless it's going to break other devices I say leave as is as it restores functionality for a good set of users.  

If it can be trimmed to work with all devices and not take away from any that's absolutely fine.  But last time somebody messed with it, they gimped the d2vzw pretty hard.  If only that was scrutinized as much as this fix.

Patch-set: 1