Update patch set 1
Patch Set 1:

Shareef, I edited the lines you speak about to remove "V4V6" and then compared to what CM currently uses for VZW APNs and now they are basically a match.  Which means it goes back to not working.

Has anybody actually applied this patch and moved from a 3G area to 4G area?  Or noticed that they don't get stuck at random times with good signal and no data connection?

I think the problem here is if you are in good 4G or don't have a VZW phone, you might not see the issue that this patch addresses.  If you are like me and are constantly inbetween 3G and 4G, you will see how great this works when you no longer have to airplane mode toggle or reboot to get a data connection again.

Patch-set: 1