Update patch set 1
Patch Set 1:

noobnl: It's time to cut the cargo cult crap.

This whole thing started because using IPV4V6 broke LTE/eHRPD handoff on d2spr.  It was politely pointed out to us that, despite what the apn configuration in telephony.db states, d2psr _doesn't_ use dual-stack on stock ROM, IP(v4) only.  This was confirmed with logs, a snippet of which is posted as the commit message for I938bf9d1a0b50ec72283fbdba3785832519f3f48 [1].  So we force Sprint apns to use "IP" only.

Here's the thing, I'm aware of no evidence that qcom can't support dual stack.  It's either a network issue, or a combination of network and the telephony stack.  Either way, I'm not convinced that the same solution necessarily applies to other carriers, particularly where evidence suggests otherwise.

Jon: Have you/can you confirm from logs what protocol stack is used in SETUP_DATA_CALL on d2vzw stock ROM?

If it's actually using IPV4V6 then absolutely we should use it here.  If it's not, something curious is going on, but we should still go with the solution that works empirically, so long as it's had sufficient testing to know it's appropriate across most/many Verizon SIDs.

[1] http://review.cyanogenmod.org/#/c/27507/

Patch-set: 1