Update patch set 1
Patch Set 1:

Of course the Browser shouldn't be remove by default but in case of the Calendar I think it makes no sense to have both installed. Especially because the dual notifications. 

When you want to make some debugging of the stock Calendar you can simply deinstall the Play Store version and nothing will happen.

Patch-set: 1