Update patch set 1
Patch Set 1:

Isn't it easier for an attacker who gains temp-root on the device to permanently disable that security feature if it's set in build.prop than if it's set in default.prop that is a part of the ramdisk?

Although I suppose the attacker could just run setprop to disable it when the device has booted.

Patch-set: 1