Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Yeah, inserting this into a much-harder-to-modify piece of the filesystem is definitely intentional. Also, we don't want the usual third-party tweaker stuff to mess with this.

Recovery can be addressed in different ways (such as disabling the auth hooks in adbd if there are no keys in /data/)

Patch-set: 1