Update patch set 1
Patch Set 1:

Old habit. We didn't default live wallpapers because not all devices supported them, so the devices that could changed the default wallpaper provider (which is a different provider from the default ImageWallpaper, which we've just reconfigured). It stuck, and we're probably going to have to deal with it soon(ish)

Patch-set: 1