Update patch set 1
Patch Set 1:

(1 comment)

Need to update jfltespr model number (SPH-L720), see inline.

Otherwise I'm OK with this.  I'll note that the union of d2/jf contributors in the list is less than the number of folks who push commits through the "D2 Team" ACL, which is fine from my perspective.

I invite these folks to participate as "advertised maintainers" if they wish, otherwise they may feel free continue to operate as they presently do, if they would prefer to not wear the maintainer hat.

That said, if you _do_ want your contributions to reflect in maintainer status and are willing to accept accountability for that, please add yourself in an inline comment.  I don't believe we're trying to be exclusive here.

Patch-set: 1