Update patch set 2
Patch Set 2:

Hey Steve,

Just wanted to let you know I compiled this, the toggle, and the added repo, but left this one out just to test:
http://review.cyanogenmod.org/#/c/47774/

When compiling it complained about something in the fastpoweron app when trying to compile /packages/apps/fastpoweron

I know you said in the functionality patch that there was some work to be done.  I just wanted to let you know incase it was caused by something else.

If I needed that 3rd patch for it even to compile then my bad.  Was just trying to test the settings button, etc first.

Patch-set: 2