Update patch set 1
Patch Set 1:

We could ask people to neutralize this in the packages they make, but we have absolutely no control over what people flash on top of CM or where it comes from. Even if the majority of packagers did this right, we'd still see broken ones in the wild (not to mention the ones already out there today); it's harmless to do it on our end, and it gets rid of the problem reliably.

Patch-set: 1