Update patch set 4
Patch Set 4:

Any power user is going to prefer adb shell over a terminal emulator. Further, all get-prebuilts is doing is curl'ing the apk from Jack's github - so even a potential argument that "not everyone uses market" falls flat. 

http://jackpal.github.com/Android-Terminal-Emulator/downloads/Term.apk

There is almost no use case for this app to be in a release-key signed user build. Userdebug/eng sure, but not user.

Patch-set: 4