Update patch set 1
Patch Set 1: Code-Review-1

Just here for testing. AFAIK the 316 can't be reached by any Android device (yet). My understanding is it's being refarmmed for LTE. So maybe it needs a bearer to define it as such. The Sprint world APN I have connected to on the other hand. But I'm unsure if it was due to MCC MNC match or the modem actually seen the APN out there. I also do not know if this APN should require a bearer or not for LTE/eHRPD.

Patch-set: 1
Reviewer: Richard Ross <314@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-1